fix(be): prevent monetary integer overflow
Migrate Order subTotal/discountAmount/taxAmount/serviceChargeAmount/totalPrice and Transaction amount/cashReceived/changeGiven and OrderItem price/discountAmount/totalPrice from INTEGER to BIGINT to support legitimate large orders (e.g. 1M * 3000 = 3B > 2.1B INT4 max).

Handle pg int8 as Number via pg.types.setTypeParser(20) for API compatibility (safe <9e15).

Add regression test bigint-large-order.test.js for 3B persistence.

+++ db/models/order_item.js
@@ -26,7 +26,7 @@ module.exports = (sequelize, DataTypes) => {
       },
       price: {
         allowNull: false,
-        type: DataTypes.INTEGER
+        type: DataTypes.BIGINT
       },
       discountType: {
         type: DataTypes.ENUM('none', 'percent', 'nominal'),
@@ -37,11 +37,11 @@ module.exports = (sequelize, DataTypes) => {
         defaultValue: 0
       },
       discountAmount: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.BIGINT,
         defaultValue: 0
       },
       totalPrice: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.BIGINT,
         defaultValue: 0
       },
       options: {
